moved endpoints to proper controller
